Multiple changes:
 - Add new `ChainProgress` enum as interface between BHW and BW.
 - Fix bug where going from the first special election to the proper election
   consensus leaves some wrong state.
 - Restructure into multiple files.
