+++ state-chain/pallets/cf-elections/src/electoral_systems/block_height_tracking.rs
@@ -11,6 +11,7 @@ use crate::{
 use cf_chains::btc::BlockNumber;
 use cf_utilities::success_threshold_from_share_count;
 use codec::{Decode, Encode};
+use consensus::{Consensus, StagedConsensus, SupermajorityConsensus, Threshold};
 use frame_support::{
 	ensure,
 	pallet_prelude::{MaybeSerializeDeserialize, Member},
@@ -18,6 +19,7 @@ use frame_support::{
 	Parameter,
 };
 use itertools::Itertools;
+use primitives::{trim_to_length, validate_vote_and_height, ChainBlocks, Header, MergeFailure};
 use scale_info::TypeInfo;
 use serde::{Deserialize, Serialize};
 use sp_std::{
@@ -25,14 +27,8 @@ use sp_std::{
 	vec::Vec,
 };
 
-#[derive(
-	Debug, Clone, PartialEq, Eq, Encode, Decode, TypeInfo, Deserialize, Serialize, Ord, PartialOrd,
-)]
-pub struct Header<BlockHash, BlockNumber> {
-	pub block_height: BlockNumber,
-	pub hash: BlockHash,
-	pub parent_hash: BlockHash,
-}
+pub mod consensus;
+pub mod primitives;
 
 #[derive(
 	Debug, Clone, PartialEq, Eq, Encode, Decode, TypeInfo, Deserialize, Serialize, Ord, PartialOrd,
@@ -41,177 +37,6 @@ pub struct BlockHeightTrackingProperties<BlockNumber> {
 	/// An election starts with a given block number,
 	/// meaning that engines have to submit all blocks they know of starting with this height.
 	pub witness_from_index: BlockNumber,
-}
-
-#[derive(
-	Debug, Clone, PartialEq, Eq, Encode, Decode, TypeInfo, Deserialize, Serialize, Ord, PartialOrd,
-)]
-
-/// Invariant:
-/// This should always be a continuous chain of block headers
-pub struct ChainBlocks<H, N> {
-	pub headers: VecDeque<Header<H, N>>,
-	pub next_height: N,
-}
-
-impl<H, N> ChainBlocks<H, N> {}
-
-pub enum ChainBlocksMergeResult<N> {
-	Extended { new_highest: N },
-	FailedMissing { range: Range<N> },
-}
-
-pub struct MergeInfo<H, N> {
-	removed: VecDeque<Header<H, N>>,
-	added: VecDeque<Header<H, N>>,
-}
-
-impl<H, N: Copy> MergeInfo<H, N> {
-	pub fn get_added_block_heights(&self) -> Option<RangeInclusive<N>> {
-		if let (Some(first), Some(last)) = (self.added.front(), self.added.back()) {
-			Some(first.block_height..=last.block_height)
-		} else {
-			None
-		}
-	}
-}
-
-enum MergeFailure<H, N> {
-	// If we get a new range of blocks, [lowest_new_block, ...], where the parent of
-	// `lowest_new_block` should, by block number, be `existing_wrong_parent`, but who's
-	// hash doesn't match with `lowest_new_block`'s parent hash.
-	ReorgWithUnknownRoot { new_block: Header<H, N>, existing_wrong_parent: Option<Header<H, N>> },
-
-	// /// This means that we have requested blocks which start higher than our last highest
-	// block, /// should not happen if everything goes well.
-	// MissingBlocks { range: Range<N>},
-	InternalError,
-}
-
-pub fn extract_common_prefix<A: Eq>(a: &mut VecDeque<A>, b: &mut VecDeque<A>) -> VecDeque<A> {
-	let mut prefix = VecDeque::new();
-
-	while a.front().is_some() && (a.front() == b.front()) {
-		prefix.push_front(a.pop_front().unwrap());
-		b.pop_front();
-	}
-
-	prefix
-}
-
-pub fn trim_to_length<A>(items: &mut VecDeque<A>, target_length: usize) -> VecDeque<A> {
-	let mut result = VecDeque::new();
-	while items.len() > target_length {
-		if let Some(front) = items.pop_front() {
-			result.push_back(front);
-		}
-	}
-	result
-}
-
-pub fn head_and_tail<A: Clone>(mut items: &VecDeque<A>) -> Option<(A, VecDeque<A>)> {
-	let items = items.clone();
-	items.clone().pop_front().map(|head| (head, items))
-}
-
-enum VoteValidationError {
-	BlockHeightsNotContinuous,
-	ParentHashMismatch,
-}
-
-impl<
-		H: Eq + Clone,
-		N: Ord
-			+ From<u32>
-			+ Add<N, Output = N>
-			+ Sub<N, Output = N>
-			+ SubAssign<N>
-			+ AddAssign<N>
-			+ Copy,
-	> ChainBlocks<H, N>
-{
-	fn validate(&self) -> Result<(), VoteValidationError> {
-		let mut required_block_height = self.next_height - 1u32.into();
-		let mut required_hash = None;
-
-		for header in self.headers.iter().rev() {
-			ensure!(
-				header.block_height == required_block_height,
-				VoteValidationError::BlockHeightsNotContinuous
-			);
-			ensure!(
-				Some(&header.hash) == required_hash.as_ref().or(Some(&header.hash)),
-				VoteValidationError::ParentHashMismatch
-			);
-
-			required_block_height -= 1u32.into();
-			required_hash = Some(header.parent_hash.clone());
-		}
-
-		Ok(())
-	}
-
-	//
-	// We have the following assumptions:
-	//
-	// Assumptions:
-	//   1. `other`: a. is well-formed (contains incrementing heights) b. is nonempty
-	// 	 2. `self`: a. is well-formed (contains incrementing heights)
-	//   3. one of the following cases holds
-	//       - case 1: `other` starts exactly after `self` ends OR self is `Default::default()`
-	//       - case 2: (`self` and `other` start at the same block) AND (self is nonempty)
-	//
-	pub fn merge(
-		&mut self,
-		other: VecDeque<Header<H, N>>,
-	) -> Result<MergeInfo<H, N>, MergeFailure<H, N>> {
-		// assumption (1b)
-		let other_head = other.front().ok_or(MergeFailure::InternalError)?;
-
-		if self.next_height == other_head.block_height ||
-			(self.next_height == 0u32.into() && self.headers.len() == 0)
-		{
-			// this is "assumption (3): case 1"
-			//
-			// This means that our new blocks start exactly after the ones we already have,
-			// so we have to append them to our existing ones. And make sure that the hash/parent
-			// hashes match.
-
-			if match self.headers.back() {
-				None => true,
-				Some(h) => other_head.parent_hash == h.hash,
-			} {
-				self.next_height += (other.len() as u32).into();
-				self.headers.append(&mut other.clone());
-				Ok(MergeInfo { removed: VecDeque::new(), added: other })
-			} else {
-				Err(MergeFailure::ReorgWithUnknownRoot {
-					new_block: other_head.clone(),
-					existing_wrong_parent: self.headers.back().cloned(),
-				})
-			}
-		} else {
-			// this is "assumption (3): case 2"
-
-			let self_head = self.headers.front().ok_or(MergeFailure::InternalError)?.clone();
-
-			if self_head.block_height == other_head.block_height {
-				// extract common prefix of headers
-				let mut self_headers = self.headers.clone();
-				let mut other_headers = other.clone();
-				let common_headers = extract_common_prefix(&mut self_headers, &mut other_headers);
-
-				// set headers to `common_headers` + `other_headers`
-				self.headers = common_headers;
-				self.headers.append(&mut other_headers.clone());
-				self.next_height = self_head.block_height + (self.headers.len() as u32).into();
-
-				Ok(MergeInfo { removed: self_headers, added: other_headers })
-			} else {
-				Err(MergeFailure::InternalError)
-			}
-		}
-	}
 }
 
 #[derive(
@@ -243,98 +68,17 @@ pub fn validate_vote<ChainBlockHash, ChainBlockNumber>(
 ) {
 }
 
-// -- abstract Consensus for computing solutions
-pub trait Consensus: Default {
-	type Vote;
-	type Result;
-	type Settings;
-	fn insert_vote(&mut self, vote: Self::Vote);
-	fn check_consensus(&self, settings: &Self::Settings) -> Option<Self::Result>;
-}
-
-struct SupermajorityConsensus<Vote: PartialEq> {
-	votes: BTreeMap<Vote, u32>,
-}
-
-struct Threshold {
-	threshold: u32,
-}
-
-impl<Vote: PartialEq> Default for SupermajorityConsensus<Vote> {
-	fn default() -> Self {
-		Self { votes: Default::default() }
-	}
-}
-
-impl<Vote: Ord + PartialEq + Clone> Consensus for SupermajorityConsensus<Vote> {
-	type Vote = Vote;
-	type Result = Vote;
-	type Settings = Threshold;
-
-	fn insert_vote(&mut self, vote: Self::Vote) {
-		if let Some(count) = self.votes.get_mut(&vote) {
-			*count += 1;
-		} else {
-			self.votes.insert(vote, 1);
-		}
-	}
-
-	fn check_consensus(&self, settings: &Self::Settings) -> Option<Self::Result> {
-		let best = self.votes.iter().last();
-
-		if let Some((best_vote, best_count)) = best {
-			if best_count >= &settings.threshold {
-				return Some(best_vote.clone());
-			}
-		}
-
-		return None;
-	}
-}
-
-// --
-struct StagedConsensus<Stage: Consensus, Index: Ord> {
-	stages: BTreeMap<Index, Stage>,
-}
-
-impl<Stage: Consensus, Index: Ord> StagedConsensus<Stage, Index> {
-	pub fn new() -> Self {
-		Self { stages: BTreeMap::new() }
-	}
-}
-
-impl<Stage: Consensus, Index: Ord> Default for StagedConsensus<Stage, Index> {
-	fn default() -> Self {
-		Self { stages: Default::default() }
-	}
-}
-
-impl<Stage: Consensus, Index: Ord + Copy> Consensus for StagedConsensus<Stage, Index> {
-	type Result = Stage::Result;
-	type Vote = (Index, Stage::Vote);
-	type Settings = Stage::Settings;
-
-	fn insert_vote(&mut self, (index, vote): Self::Vote) {
-		if let Some(stage) = self.stages.get_mut(&index) {
-			stage.insert_vote(vote)
-		} else {
-			let mut stage = Stage::default();
-			stage.insert_vote(vote);
-			self.stages.insert(index, stage);
-		}
-	}
-
-	fn check_consensus(&self, settings: &Self::Settings) -> Option<Self::Result> {
-		// we check all stages starting with the highest index,
-		// the first one that has consensus wins
-		for (_, stage) in self.stages.iter().rev() {
-			if let Some(result) = stage.check_consensus(settings) {
-				return Some(result);
-			}
-		}
-
-		None
-	}
+#[derive(Debug, Clone, PartialEq, Eq, Encode, Decode, TypeInfo, Deserialize, Serialize)]
+pub enum ChainProgress<ChainBlockNumber> {
+	// Block witnesser will discard any elections that were started for this range and start them
+	// again since we've detected a reorg
+	Reorg(RangeInclusive<ChainBlockNumber>),
+	// the chain is just progressing as a normal chain of hashes
+	Continuous(RangeInclusive<ChainBlockNumber>),
+	// there was no update to the witnessed block headers
+	None(ChainBlockNumber),
+	// We are starting up and don't have consensus on a block number yet
+	WaitingForFirstConsensus,
 }
 
 // -- electoral system
@@ -372,7 +116,7 @@ impl<
 	type OnFinalizeContext = ();
 
 	// new block to query for the block witnesser
-	type OnFinalizeReturn = Option<RangeInclusive<ChainBlockNumber>>;
+	type OnFinalizeReturn = ChainProgress<ChainBlockNumber>;
 
 	fn generate_vote_properties(
 		_election_identifier: ElectionIdentifier<Self::ElectionIdentifierExtra>,
@@ -422,7 +166,7 @@ impl<
 								);
 
 								Ok((
-									merge_info.get_added_block_heights(),
+									merge_info.into_chain_progress().unwrap(),
 									unsynchronised_state.headers.next_height,
 								))
 							},
@@ -431,9 +175,17 @@ impl<
 								existing_wrong_parent,
 							}) => {
 								log::info!("detected a reorg: got block {new_block:?} whose parent hash does not match the parent block we have recorded: {existing_wrong_parent:?}");
-								Ok((None, unsynchronised_state.headers.next_height))
+								Ok((
+									unsynchronised_state
+										.headers
+										.current_state_as_no_chain_progress(),
+									unsynchronised_state.headers.next_height,
+								))
 							},
-							Err(MergeFailure::InternalError) => Err(CorruptStorageError::new()),
+							Err(MergeFailure::InternalError(reason)) => {
+								log::error!("internal error in block height tracker: {reason}");
+								Err(CorruptStorageError::new())
+							},
 						}
 					})?;
 
@@ -445,7 +197,9 @@ impl<
 
 				Ok(block_witnesser_range)
 			} else {
-				Ok(None)
+				Ok(ElectoralAccess::unsynchronised_state()?
+					.headers
+					.current_state_as_no_chain_progress())
 			}
 		} else {
 			log::info!("Starting new election with index 0 because no elections exist");
@@ -457,7 +211,7 @@ impl<
 				witness_from_index: 0u32.into(),
 			};
 			ElectoralAccess::new_election((), properties, ())?;
-			Ok(None)
+			Ok(ChainProgress::WaitingForFirstConsensus)
 		}
 	}
 
@@ -490,6 +244,10 @@ impl<
 					let mut headers = VecDeque::new();
 					headers.push_back(result);
 					headers
+				})
+				.map(|result| {
+					log::info!("block_height: initial consensus: {result:?}");
+					result
 				}))
 		} else {
 			// This is the actual consensus finding, once the engine is running
@@ -498,6 +256,12 @@ impl<
 				StagedConsensus::new();
 
 			for mut vote in consensus_votes.active_votes() {
+				// ensure that the vote is valid
+				if let Err(err) = validate_vote_and_height(properties.witness_from_index, &vote) {
+					log::warn!("received invalid vote: {err:?} ");
+					continue;
+				}
+
 				// we count a given vote as multiple votes for all nonempty subchains
 				while vote.len() > 0 {
 					consensus.insert_vote((vote.len(), vote.clone()));
@@ -505,9 +269,19 @@ impl<
 				}
 			}
 
-			Ok(consensus.check_consensus(&Threshold {
-				threshold: success_threshold_from_share_count(num_authorities),
-			}))
+			Ok(consensus
+				.check_consensus(&Threshold {
+					threshold: success_threshold_from_share_count(num_authorities),
+				})
+				.map(|result| {
+					log::info!(
+						"(witness_from: {:?}): successful consensus for ranges: {:?}..={:?}",
+						properties.witness_from_index,
+						result.front(),
+						result.back()
+					);
+					result
+				}))
 		}
 	}
 }
